pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dify-admin"
version = "0.1.0"
description = "Dify Console API client library and CLI — manage apps, knowledge bases, and auth without the GUI"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
    "mypy>=1.0",
]
mcp = [
    "mcp>=1.0",
]
db = [
    "psycopg2-binary>=2.9",
]

[project.scripts]
dify-admin = "dify_admin.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
